Initial commit
Add four projects for each phase of the exercise (phase 5 has not been
developed yet). 
Phase 1 is the implementation of the game without any pattern being
applied. 
Phase 2 is the implemetation of a game timer using the observer and
adapter patterns.
Phase 3 is the implementation of themes for the images of the game using
the abstract factory pattern.
Phase 4 is the implementation of a change that allow easily defining new
difficulty levels using the template method pattern.